ACMS-4504: Move FunctionalJavascript test to FunctionalJavascript directory.
